[#83] Reuse data-migration test fixtures
- Moves backfill test to tests/data_migrations, drops private engine
- Tests added: no
- Docs updated: no
